+++ src/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+- Added hub edit ability to hubs-dialog.
 - seqs dialog replaced with hubs dialog which will become a hub centric dialog.
 - Added hub store for storing user data in indexdb.
 - genesets now part of trackable history.